refactor: move inline lynx tests to proper test file
Move test_read_url() and test_search() from production code
(_browser_lynx.py) to tests/test_browser_lynx.py with proper
@slow markers and lynx-availability skip conditions.
