+++ tests/test_browser_lynx.py
@@ -1,8 +1,12 @@
-"""Tests for lynx browser backend security."""
+"""Tests for lynx browser backend."""
+
+import shutil
 
 import pytest
 
-from gptme.tools._browser_lynx import _validate_url_scheme
+from gptme.tools._browser_lynx import _validate_url_scheme, read_url, search
+
+lynx_available = shutil.which("lynx") is not None
 
 
 def test_url_scheme_validation():
@@ -21,3 +25,21 @@ def test_url_scheme_validation():
 
     with pytest.raises(ValueError, match="not allowed"):
         _validate_url_scheme("javascript:alert(1)")
+
+
+@pytest.mark.slow
+@pytest.mark.skipif(not lynx_available, reason="lynx not installed")
+def test_read_url():
+    """Test reading URLs with lynx backend."""
+    content = read_url("https://gptme.org/")
+    assert "Getting Started" in content
+    content = read_url("https://github.com/gptme/gptme/issues/205")
+    assert "lynx-backed browser tool" in content
+
+
+@pytest.mark.slow
+@pytest.mark.skipif(not lynx_available, reason="lynx not installed")
+def test_search():
+    """Test search with lynx backend."""
+    result = search("Erik Bjäreholt", "duckduckgo")
+    assert "erik.bjareholt.com" in result
